Add tests to pgclient for common data types. Fix protocol so that binary and time.Time types work.

+++ snapshotserver/snapshot.go
@@ -100,7 +100,7 @@ func GetTenantSnapshotData(tenantId []string, conn *pgclient.PgConnection) (b []
 			for ind, y := range x {
 				scv := &common.ColumnVal{
 					Value: y,
-					Type:  ci[ind].Type,
+					Type:  int32(ci[ind].Type),
 				}
 				srvItem[ci[ind].Name] = scv
 			}
